requirements.txt
aiogram
python-dotenv
sqlalchemy
Flask
